fix for warnings
git-svn-id: http://svn.miranda-ng.org/main/trunk@13902 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Tox/src/tox_proto.h
@@ -277,7 +277,7 @@ private:
 
 	static bool IsFileExists(std::tstring path);
 
-	MEVENT AddEventToDb(MCONTACT hContact, WORD type, DWORD timestamp, DWORD flags, PBYTE pBlob, DWORD cbBlob);
+	MEVENT AddEventToDb(MCONTACT hContact, WORD type, DWORD timestamp, DWORD flags, PBYTE pBlob, size_t cbBlob);
 
 	template<INT_PTR(__cdecl CToxProto::*Service)(WPARAM, LPARAM)>
 	static INT_PTR __cdecl GlobalService(WPARAM wParam, LPARAM lParam)
